매개변수 타입 변경
int -> long

+++ javaProject/MyProject/src/quiz02/Calculator.java
@@ -11,7 +11,7 @@ public class Calculator {
 		System.out.println(num1 + " - " + num2 + " = " + (num1 - num2));
 	}
 	
-	public static void mul (int num1, int num2) {
+	public static void mul (long num1, long num2) {
 		System.out.println(num1 + " * " + num2 + " = " + (num1 * num2));
 	}
